+++ Raspi/Test/serial_test.py
@@ -2,7 +2,7 @@ import argparse
 import os
 import paho.mqtt.client as mqtt
 
-from test_data_generator import send_csv_data, send_fake_data
+from das_data_generator import send_csv_data, send_fake_data
 
 # Arguments
 parser = argparse.ArgumentParser(description='DAS serial test script', add_help=True)
